requirements.txt
# Core dependencies matching specification requirements
torch>=1.13.0
torchaudio>=0.13.0
torchvision>=0.14.0
transformers>=4.25.0
datasets>=2.8.0
librosa>=0.10.0
soundfile>=0.12.0
numpy>=1.21.0
scipy>=1.9.0
matplotlib>=3.6.0
tqdm>=4.64.0

# Audio processing and feature extraction
speechbrain>=0.5.14
pydub>=0.25.1
webrtcvad>=2.0.10

# Text processing
phonemizer>=3.2.0
unidecode>=1.3.0
g2p-en>=2.1.0

# Training and optimization
wandb>=0.13.0
tensorboard>=2.11.0
scikit-learn>=1.1.0
hydra-core>=1.3.0

# Web API and deployment
fastapi>=0.88.0
uvicorn>=0.20.0
python-multipart>=0.0.5
websockets>=10.4

# Utility and monitoring
pyyaml>=6.0
requests>=2.28.0
psutil>=5.9.0
GPUtil>=1.4.0

# Testing
pytest>=7.2.0
pytest-cov>=4.0.0